NTM initial state bias values, cross-entropy fixed
After leaving the training running all night, noticed that read/write address vectors contained elements that were ~1/N.

Added method that returns a bias value for the NTM's initial state. Specifically, the initial read/write addresses are set to a one-hot encoded vector with the zeroth element being the hot element. The goal is to allow the 'seed' the network into learning how to generate addresses by providing it with the correct address first.

Also modified the loss function so that cross-entropy is only calculated over the k bits that are supposed to be written to the memory matrix. Before cross-entropy was computed across the k+1 bits comprising the entire instruction (the extra bit is a delimiter that indicates that the network should begin regurgitating stored values.

Training now takes places with 1 - 20 instructions per batch. Similar to the paper, the goal is to have the network learn a 'general' copy method so that it can perform the copy