Modified loss function, added gradient clipping
Changed loss function to sigmoid_cross_entropy_with_logits

Added gradient clipping (gradients are clipped between -10 and 10, just like Graves' implementation).

Read a few synopses of other implementations - they had to let the training run for 35k - 75k before the copy task worked correctly.

I'm standing by my understanding that the memory matrix is part of the recurrent hidden state. According to the paper, the memory matrix is reset at the start of each training batch. This probably wouldn't be the case if the matrix was an integrated part of the training process.