Found errors, getting 'nan' sometimes
Not currently working.

Changes:
1. The working program was using incorrect input to train the NTM; There should have been no information duplication in the training vector. Training data now contains:
Bit pattern, indicator bit, blank
2. Also noticed that the read/write heads were being concatenated together incorrectly before being sent to the NTM. 

Errors:
After several thousand training iterations the network will throw a 'nan', even when the learning rate is set to 1e-4. It is odd that this particular type of error persists; the only to coax numpy into generating a 'nan' value is when operations on a real number result in a complex number. It is most likely the case that the sharpening operation is causing this.
However, it's odd to notice that when the network's output is examined (after receiving a 'nan'), all values in the read/write heads are set to 'nan', which could indicate that something awful is happening between the input layer and the LSTM.

N