Still Testing
Was able to get to 50% training error using a learning rate of 1e-4, but this required 50k iterations with batches of size 64.

Might consider decreasing the length of the copy sequence since the ratio of sequence length to number of memory locations is so high. In Graves' paper they trained on sequences of length 20, but had storage capacity for 128 sequences.