Fixed last nan error!
The recurrent 'nan' error was due to the sharpening operation becoming too... sharp. Raising a vector whose components were < 1 to a large number yielded a vector whose components were ~0, leading to a 0/0 case.

Instead of adding a small offset to the denominator of the normalized sharpening function, the sharpening constant had its output modified from a softplus to a sigmoid with a cap of 50 and an offset of 1. This heuristic allowed the NTM to load a previously saved model with ~23% error and decrease that error through further training.

Adding an offset to the denominator of the sharpening operation caused the error of the previously saved model to shoot back up to ~69%.

The 'utils.py' file was also added; these methods need to be integrated with the NTM training procedure. Also need to grab the weights used for reading/writing during validation and save those as an image.

The 'models' folder contains the most recently saved model with ~18% error (obtained with a learnin