Added 'associative recall' task
Split the training methods away from the 'ntm.py' class. Created a method for generating associative recall training batches, and trained on the associative recall task.

Model gets around 5e-3 cross-entropy error, and is able to produce the correct sequence for a series of 3 random 6-bit vectors, and reproduces the correct sequence with visible error for a series of 5 random 6-bit vectors.