Added priority sort batch generator
A training batch generator for priority sort was added to the train.py file. Also a small, incorrect version of training the NTM to perform priority sort was added. This version still uses only one read/write head.

In the paper they report very non-zero asymptotic cross-entropy errors on the priority sort task, and I think this is because they ask the NTM to store the data, then immediately ask for the sorted data once the raw data has been presented. This seems to force the NTM to learn a way to store the data somewhat sequentially as it's presented. Instead, it would be better to present the data, wait a few steps (allow the NTM to perform sorting operations after receiving the data), and then ask for the sorted data. This is what the current priority sort batch generator does; it waits nlogn timesteps (where n = number of sortable sequences) before asking for the sorted data. With one read/write head, this doesn't seem to work.

The next goal should be to mimic the