More multi-head changes, priority sort changed
The NTM Cell continues to be modified to work with multiple heads, some edits have been made to the NTM program to accommodate the NTM Cell returning a list of reads (these are far from finished).

Changed the priority sort batch generator/trainer: the batch generator now pulls a priority from a pre-defined set of numbers (n divisions of the interval [0, 1]). The trainer only presents a finite number of items to the network (in this case, 20). The trainer still uses an NTM with just one read/write head.